Added the AuthUser functionality
Added the AuthUser functionality, which enables users to sucessfully login if both email and password is correct. The method checks password by using the salt and hashed password, and also handles if either password or email is not found/matching the data in the database.
OBS: Comments still need to be added.